Using decimal instead of double
Use ToString to prevent loss of precision

+++ ExpressionEvaluator/Procedures/Functions/Sum.cs
@@ -9,7 +9,7 @@ namespace ExpressionEvaluator.Procedures.Functions
         {
             _name2 = "Sum";
             OperandList = x => {
-                double sum = 0;
+                decimal sum = 0;
                 foreach (var i in x) {
                     sum += i;
                 }
